Exact definition of M0
To quote the NMSOP, chapter 3, p.125, eq. 3.68:

Mw = (log M0 – 9.1)/1.5 = (2/3) (log M0 – 9.1)       

Formula  (3.68)  is  now  the  IASPEI  (2005  and  2013)  accepted  standard  form of  writing  the moment  magnitude  formula.  It  avoids  rounding  errors  and  an  ambiguity  that  arises  if  multiplication is performed prior to subtraction, which in a certain percentage of cases leads to  Mw  being  different  according  to  whether  the  moment  is  expressed  in  CGS  or  SI  units  
(Utsu, 2002, and  IS 3.3).