Added support for output SAC headers with geocentric coordinate system
Instaseis now accepts "sacheader" parameter with the default value of "geodetic" and optional value of "geocentric" to control how an output SAC file header is populated. If the sacheader parameter is not provided or is set to "geodetic", there will be no change in the output SAC headers. However, if the parameter is set to "geocentric", the output SAC files header will be populated using geocentric coordinates. This option is intended as a support for the Mars models.